+++ dashboard/status/model.py
